pkg/executor: move recover tests to isolated suite
