+++ pkg/executor/test/recovertest/BUILD.bazel
@@ -11,14 +11,23 @@ go_test(
     shard_count = 2,
     deps = [
         "//pkg/config",
+        "//pkg/config/kerneltype",
         "//pkg/ddl/util",
+        "//pkg/errno",
+        "//pkg/infoschema",
         "//pkg/kv",
         "//pkg/meta/autoid",
+        "//pkg/meta/model",
+        "//pkg/parser/auth",
+        "//pkg/sessionctx/variable",
         "//pkg/store/mockstore",
         "//pkg/testkit",
         "//pkg/testkit/testfailpoint",
+        "//pkg/types",
         "//pkg/util",
+        "//pkg/util/dbterror",
         "//pkg/util/gcutil",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//oracle",
         "@com_github_tikv_client_go_v2//tikv",
